requirements.txt
fastmcp>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
